pyproject.toml
[project]
authors = [
    {name = "some name", email = "some@gmail.com"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Project - Task Manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "rollbar>=1.3.0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "ruff>=0.11.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]